Store encoded password in EncodedResult object
